+++ graphrag/query/cli.py
@@ -14,70 +14,10 @@ from graphrag.config import (
     create_graphrag_config,
 )
 from graphrag.index.progress import PrintProgressReporter
-from graphrag.model.entity import Entity
-from graphrag.query.input.loaders.dfs import (
-    store_entity_semantic_embeddings,
-)
-from graphrag.vector_stores import VectorStoreFactory, VectorStoreType
-from graphrag.vector_stores.lancedb import LanceDBVectorStore
 
-from .factories import get_global_search_engine, get_local_search_engine
-from .indexer_adapters import (
-    read_indexer_covariates,
-    read_indexer_entities,
-    read_indexer_relationships,
-    read_indexer_reports,
-    read_indexer_text_units,
-)
+from . import api
 
 reporter = PrintProgressReporter("")
-
-
-def __get_embedding_description_store(
-    entities: list[Entity],
-    vector_store_type: str = VectorStoreType.LanceDB,
-    config_args: dict | None = None,
-):
-    """Get the embedding description store."""
-    if not config_args:
-        config_args = {}
-
-    collection_name = config_args.get(
-        "query_collection_name", "entity_description_embeddings"
-    )
-    config_args.update({"collection_name": collection_name})
-    description_embedding_store = VectorStoreFactory.get_vector_store(
-        vector_store_type=vector_store_type, kwargs=config_args
-    )
-
-    description_embedding_store.connect(**config_args)
-
-    if config_args.get("overwrite", False):
-        # this step assumps the embeddings where originally stored in a file rather
-        # than a vector database
-
-        # dump embeddings from the entities list to the description_embedding_store
-        store_entity_semantic_embeddings(
-            entities=entities, vectorstore=description_embedding_store
-        )
-    else:
-        # load description embeddings to an in-memory lancedb vectorstore
-        # to connect to a remote db, specify url and port values.
-        description_embedding_store = LanceDBVectorStore(
-            collection_name=collection_name
-        )
-        description_embedding_store.connect(
-            db_uri=config_args.get("db_uri", "./lancedb")
-        )
-
-        # load data from an existing table
-        description_embedding_store.document_collection = (
-            description_embedding_store.db_connection.open_table(
-                description_embedding_store.collection_name
-            )
-        )
-
-    return description_embedding_store
 
 
 def run_global_search(
@@ -88,7 +28,10 @@ def run_global_search(
     response_type: str,
     query: str,
 ):
-    """Run a global search with the given query."""
+    """Perform a global search with a given query.
+
+    Loads index files required for global search and calls the Query API.
+    """
     data_dir, root_dir, config = _configure_paths_and_settings(
         data_dir, root_dir, config_dir
     )
@@ -104,21 +47,15 @@ def run_global_search(
         data_path / "create_final_community_reports.parquet"
     )
 
-    reports = read_indexer_reports(
-        final_community_reports, final_nodes, community_level
-    )
-    entities = read_indexer_entities(final_nodes, final_entities, community_level)
-    search_engine = get_global_search_engine(
-        config,
-        reports=reports,
-        entities=entities,
+    return api.global_search(
+        config=config,
+        final_nodes=final_nodes,
+        final_entities=final_entities,
+        final_community_reports=final_community_reports,
+        community_level=community_level,
         response_type=response_type,
+        query=query,
     )
-
-    result = search_engine.search(query=query)
-
-    reporter.success(f"Global Search Response: {result.response}")
-    return result.response
 
 
 def run_local_search(
@@ -129,7 +66,10 @@ def run_local_search(
     response_type: str,
     query: str,
 ):
-    """Run a local search with the given query."""
+    """Perform a local search with a given query.
+
+    Loads index files required for local search and calls the Query API.
+    """
     data_dir, root_dir, config = _configure_paths_and_settings(
         data_dir, root_dir, config_dir
     )
@@ -151,41 +91,19 @@ def run_local_search(
         else None
     )
 
-    vector_store_args = (
-        config.embeddings.vector_store if config.embeddings.vector_store else {}
-    )
-
-    reporter.info(f"Vector Store Args: {vector_store_args}")
-    vector_store_type = vector_store_args.get("type", VectorStoreType.LanceDB)
-
-    entities = read_indexer_entities(final_nodes, final_entities, community_level)
-    description_embedding_store = __get_embedding_description_store(
-        entities=entities,
-        vector_store_type=vector_store_type,
-        config_args=vector_store_args,
-    )
-    covariates = (
-        read_indexer_covariates(final_covariates)
-        if final_covariates is not None
-        else []
-    )
-
-    search_engine = get_local_search_engine(
-        config,
-        reports=read_indexer_reports(
-            final_community_reports, final_nodes, community_level
-        ),
-        text_units=read_indexer_text_units(final_text_units),
-        entities=entities,
-        relationships=read_indexer_relationships(final_relationships),
-        covariates={"claims": covariates},
-        description_embedding_store=description_embedding_store,
+    # call the Query API
+    return api.local_search(
+        config=config,
+        final_nodes=final_nodes,
+        final_entities=final_entities,
+        final_community_reports=final_community_reports,
+        final_text_units=final_text_units,
+        final_relationships=final_relationships,
+        final_covariates=final_covariates,
+        community_level=community_level,
         response_type=response_type,
+        query=query,
     )
-
-    result = search_engine.search(query=query)
-    reporter.success(f"Local Search Response: {result.response}")
-    return result.response
 
 
 def _configure_paths_and_settings(
